refactor(music): move `IconPrefixedLabel` to central module
